# docs/openapi.yml
openapi: 3.0.0
info:
  title: MIMIR API
  version: 0.1.0
  description: The API documentation for the MIMIR game network.
  license:
    name: MIT
    url: 'https://opensource.org/licenses/mit-license.php'
servers:
  - url: 'api.mimir.nakasar.me'
    description: 'Production server'
  - url: 'http://localhost{port}'
    variables:
      port:
        default: ':5010'
        description: 'Port for development server, leave blank for 80'
    description: 'Local development server'
...
components:
...
  schemas:
    GameSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        summary:
          type: string
        thumbnail:
          type: string
        note:
          type: number
...
    PlayerSummary:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
...
    SessionSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        date:
          type: string
        location:
          type: string
        status:
          $ref: "#/components/x-enums/SessionStatus"
        description:
          type: string
        participants:
          type: number
...
paths:
  '/players':
    get:
      operationId: List players
      tags:
        - player
      responses:
        '200':
          description: List of players.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerSummary'
        '401':
          description: Missing authentication.
        '403':
          description: Invalid authentication or missing authorization.
    post:
      operationId: Create player*
      tags:
        - player
      responses:
        '201':
          description: Player created.
        '401':
          description: Missing authentication.
        '403':
          description: Invalid authentication or missing authorization.
        '409':
          description: Player displayName not available or already registered.
...
  '/games':
    get:
      operationId: List games
      tags:
        - game
      responses:
        '200':
          description: List of games.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSummary"
        '401':
          description: Missing authentication.
        '403':
          description: Invalid authentication or missing authorization.
    post:
      operationId: Create game
      tags:
        - game
      responses:
        '201':
          description: Game created.
        '401':
          description: Missing authentication.
        '403':
          description: Invalid authentication or missing authorization.
        '409':
          description: Game already in database.
...
  '/sessions':
    get:
      operationId: List sessions
      tags:
        - session
      responses:
        '200':
          description: List of sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionSummary"
        '401':
          description: Missing authentication.
        '403':
          description: Invalid authentication or missing authorization.
    post:
      operationId: Create session
      tags:
        - session
      responses:
        '201':
          description: Session created.
        '401':
          description: Missing authentication.
        '403':
          description: Invalid authentication or missing authorization.
...